webApp/requirements.txt
absl-py==0.7.1
astor==0.7.1
Click==7.0
Flask==1.0.2
gast==0.2.2
grpcio==1.19.0
gunicorn==19.9.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.1.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.2
opencv-python==4.0.0.21
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
PyYAML==5.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.1
